refactor: move migrations code to libs
